# Put filewatcher in a worker, for better stability and performance