--- a/package.json
+++ b/package.json
@@ -30,7 +30,7 @@
     "deasync": "^0.1.12",
     "dotenv": "^5.0.0",
     "filesize": "^3.6.0",
-    "fswatcher-child": "^0.0.5",
+    "fswatcher-child": "^1.0.0",
     "get-port": "^3.2.0",
     "glob": "^7.1.2",
     "grapheme-breaker": "^0.3.2",