--- a/src/Watcher.js
+++ b/src/Watcher.js
@@ -1,4 +1,4 @@
-const {FSWatcher} = require('chokidar');
+const FSWatcher = require('fswatcher-child');
 const Path = require('path');
 
 /**
@@ -17,16 +17,7 @@ class Watcher {
     });
 
     this.watchedDirectories = new Map();
-
-    // Only close the watcher after the ready event is emitted
-    this.ready = false;
     this.stopped = false;
-    this.watcher.once('ready', () => {
-      this.ready = true;
-      if (this.stopped) {
-        this.watcher.close();
-      }
-    });
   }
 
   /**
@@ -125,13 +116,18 @@ class Watcher {
   }
 
   /**
+   * Add an event handler
+   */
+  once(event, callback) {
+    this.watcher.once(event, callback);
+  }
+
+  /**
    * Stop watching all paths
    */
   stop() {
     this.stopped = true;
-    if (this.ready) {
-      this.watcher.close();
-    }
+    this.watcher.close();
   }
 }