This PR puts FSWatcher into a worker, this helps to add extra stability.

Fixes:
* Segfault error inside fsevents
* Freezing on windows
* Significant performance boost in watch mode

Some tests I've done on performance:

| Description   | With current master   | with fswatcher-child     |
| --- |---                    | ---                      |
| Very small app without cache (On Mac OS X High Sierra)   | 8.8secs               | 4secs                    |
| Small app with a lot of requires and import without cache (On windows 10) | 22secs | 7secs |

`fswatcher-child`: https://github.com/DeMoorJasper/fswatcher-child

Closes #1342 Closes #1331 Closes #1394 Closes #1357 Closes #1308