@devongovett 2018-05-16
hmm interesting idea. what do you think about using our existing worker infra in parcel for this?