@DeMoorJasper 2018-05-16
@devongovett Thought it would be a better idea to make it lightweight.
The main goal was to sandbox chokidar from parcel to prevent errors and crashes. It doesn’t need multiple workers and localworker in this case wouldn’t prevent the crashes

<sub>Sent with <a href="http://githawk.com">GitHawk</a></sub>