@DeMoorJasper 2018-05-18
@devongovett I've added some benchmarks, as I've just discovered after running the inspector on a project that it also boosts dev builds on non windows machines quite a lot